% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtou.R
\name{shrinkToMinimum}
\alias{shrinkToMinimum}
\title{length of the shortest minimal unique subsequence covering each nucleotide}
\usage{
shrinkToMinimum(dtou)
}
\arguments{
\item{dtou}{A list of numeric vectors generated by \code{dtou}.}
}
\description{
\code{shrinkToMinimum} expects the output of \code{dtou} as its only input
}
\details{
For each genetic sequence in the input parameter \code{dtou}, the function
\code{shrinkToMinimum} returns the length of the
\emph{minimal unique subsequence} covering that nucleotide.

the \emph{minimal unique subsequence} has the property that shortening the
sequence, either by removing the rightmost nucleotide, or by removing
the leftmost nucleotide, results in a sub-sequence that occurs (either as-is,
or in the reverse complement) somewhere else in the set of sequences
being analyzed.
}
\examples{
test1<-dtou(c("AAAAACCCGACTGGGCTCA","ACCT"),rc=TRUE)
shrinkToMinimum(test1)

}
